+++ lib/src/home/controller/taskcards_controller.dart
@@ -79,7 +79,8 @@ class _TaskCardsState extends State<TaskCard> {
                 builder: (_, snapshot) => _futureBuilder(snapshot)),
           ),
           Padding(
-            padding: const EdgeInsets.only(top: 20, right: 30), //(top: 20, right: 25),
+            padding: const EdgeInsets.only(
+                top: 20, right: 30), //(top: 20, right: 25),
             child: Align(
               alignment: Alignment.topRight,
               child: InkWell(
@@ -131,9 +132,9 @@ class PickImage {
 
     ///todo To be removed
     String name;
-    if(con.database){
+    if (con.database) {
       name = con.module['name'];
-    }else{
+    } else {
       name = con.moduleName;
     }