+++ lib/src/home/view/screens/ios/submodules_ios.dart
@@ -8,5 +8,4 @@ import 'package:playhouse/src/view.dart';
 
 import 'package:playhouse/src/controller.dart';
 
-class SubmodulesListiOS extends SubmodulesListAndroid {
-}
+class SubmodulesListiOS extends SubmodulesListAndroid {}