+++ lib/src/view.dart
@@ -95,4 +95,3 @@ export 'package:auto_size_text/auto_size_text.dart';
 
 /// Crop images
 export 'package:crop/crop.dart';
-